requirements.txt
antlr4-python3-runtime==4.8
cycler==0.10.0
hydra-core==1.0.6
importlib-resources==5.1.2
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
omegaconf==2.0.6
Pillow==8.2.0
PyAudio==0.2.11
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
scipy==1.6.3
six==1.15.0
typing-extensions==3.10.0.0
